refactor(nyz): refactor collector name
