+++ ding/scripts/dijob-qbert.yaml
@@ -108,7 +108,7 @@ spec:
                 learner=dict(type='base', import_names=['ding.worker.learner.base_learner']),
                 collector=dict(
                     type='zergling',
-                    import_names=['ding.worker.collector.zergling_collector'],
+                    import_names=['ding.worker.collector.zergling_parallel_collector'],
                 ),
                 commander=dict(
                     type='solo',
