+++ ding/worker/collector/__init__.py
@@ -1,14 +1,15 @@
 # serial
 from .base_serial_collector import ISerialCollector, create_serial_collector, get_serial_collector_cls, \
     to_tensor_transitions
-from .sample_serial_collector import SampleCollector
-from .episode_serial_collector import EpisodeCollector
-from .episode_one_vs_one_serial_collector import Episode1v1Collector
+from .sample_serial_collector import SampleSerialCollector
+from .episode_serial_collector import EpisodeSerialCollector
+from .battle_episode_serial_collector import BattleEpisodeSerialCollector
 
 from .base_serial_evaluator import ISerialEvaluator, VectorEvalMonitor
 from .interaction_serial_evaluator import InteractionSerialEvaluator
 from .battle_interaction_serial_evaluator import BattleInteractionSerialEvaluator
 # parallel
-from .base_parallel_collector import BaseCollector, create_parallel_collector, get_parallel_collector_cls
-from .zergling_collector import ZerglingCollector
+from .base_parallel_collector import BaseParallelCollector, create_parallel_collector, get_parallel_collector_cls
+from .zergling_parallel_collector import ZerglingParallelCollector
+from .marine_parallel_collector import MarineParallelCollector
 from .comm import BaseCommCollector, FlaskFileSystemCollector, create_comm_collector, NaiveCollector
