+++ ding/worker/collector/tests/test_episode_serial_collector.py
@@ -1,5 +1,5 @@
 import pytest
-from ding.worker import EpisodeCollector
+from ding.worker import EpisodeSerialCollector
 from ding.envs import BaseEnvManager, SyncSubprocessEnvManager, AsyncSubprocessEnvManager
 from ding.policy import DQNPolicy
 from ding.model import DQN
@@ -13,7 +13,7 @@ def test_collect(env_manager_type):
     env.seed(0)
     model = DQN(obs_shape=4, action_shape=1)
     policy = DQNPolicy(DQNPolicy.default_config(), model=model).collect_mode
-    collector = EpisodeCollector(EpisodeCollector.default_config(), env, policy)
+    collector = EpisodeSerialCollector(EpisodeSerialCollector.default_config(), env, policy)
 
     collected_episode = collector.collect(
         n_episode=18, train_iter=collector._collect_print_freq, policy_kwargs={'eps': 0.5}
