+++ dizoo/classic_control/bitflip/entry/bitflip_dqn_main.py
@@ -5,7 +5,7 @@ from easydict import EasyDict
 from functools import partial
 
 from ding.config import compile_config
-from ding.worker import BaseLearner, EpisodeCollector, InteractionSerialEvaluator, EpisodeReplayBuffer
+from ding.worker import BaseLearner, EpisodeSerialCollector, InteractionSerialEvaluator, EpisodeReplayBuffer
 from ding.envs import BaseEnvManager, DingEnvWrapper
 from ding.policy import DQNPolicy
 from ding.model import DQN
@@ -22,7 +22,7 @@ def main(cfg, seed=0, max_iterations=int(1e8)):
         BaseEnvManager,
         DQNPolicy,
         BaseLearner,
-        EpisodeCollector,
+        EpisodeSerialCollector,
         InteractionSerialEvaluator,
         EpisodeReplayBuffer,
         save_cfg=True
@@ -47,7 +47,7 @@ def main(cfg, seed=0, max_iterations=int(1e8)):
     # Set up collection, training and evaluation utilities
     tb_logger = SummaryWriter(os.path.join('./{}/log/'.format(cfg.exp_name), 'serial'))
     learner = BaseLearner(cfg.policy.learn.learner, policy.learn_mode, tb_logger, exp_name=cfg.exp_name)
-    collector = EpisodeCollector(
+    collector = EpisodeSerialCollector(
         cfg.policy.collect.collector, collector_env, policy.collect_mode, tb_logger, exp_name=cfg.exp_name
     )
     evaluator = InteractionSerialEvaluator(
