+++ dizoo/gfootball/entry/parallel/gfootball_ppo_parallel_config.py
@@ -67,8 +67,8 @@ gfootball_ppo_create_config = dict(
     policy=dict(type='ppo_lstm_command', import_names=['dizoo.gfootball.policy.ppo_lstm']),
     learner=dict(type='base', import_names=['ding.worker.learner.base_learner']),
     collector=dict(
-        type='one_vs_one',
-        import_names=['ding.worker.collector.one_vs_one_collector'],
+        type='marine',
+        import_names=['ding.worker.collector.marine_parallel_collector'],
     ),
     commander=dict(
         type='one_vs_one',
